add further options for "fill"
now shaded areas generated with the "fill" option on line objects
accept the additional parameters:
"above_min" and "below_max".
this enables thresholds relative to the min and max values of a
line, in addition to absloute thresholds accessible with
"above" and "below".
this helps to define the shaded areas independent of gauge values.
for example, plotting the trapping potential without any gauge and
fill = ["below_max", -1]
results in the same shaded trapping region as
fill = ["below", -1] with gauge = "max"